src/index_monitor/requirements.txt
altair==5.3.0
attrs==23.2.0
azure-core==1.30.2
azure-identity==1.17.1       
azure-keyvault-secrets==4.8.0
azure-storage-blob==12.21.0  
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2    
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
idna==3.7
isodate==0.6.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msal==1.30.0
msal-extensions==1.2.0
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
portalocker==2.10.1
protobuf==5.27.2
#psycopg2==2.9.9
pyarrow==16.1.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.8.0
pyodbc==5.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
#pywin32==306
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rich==13.7.1
rpds-py==0.19.0
schedule==1.2.2
rpds-py==0.19.0
schedule==1.2.2
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.31
streamlit==1.36.0
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
toolz==0.12.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1